Allow download to be triggered remotely
If you want the download to be triggered e.g. by clicking a button or
some other specified area in some custom-rendered column instead of by
listening for clicks to the whole column, you won't get a GridClickEvent
from that click, but the renderer has other ways to determine the column
and row indexes by each rendered cell.